+++ doc/release_notes.rst
@@ -7,6 +7,7 @@ Release notes
 * xpxspec refactored, with most of the code being moved to evt.fitting.py
   (issue #92).
 * Some specific refactoring.
+* Equipopulated binning refactored (issue #93).
 
 
 *ximpol (0.36.1) - Sat, 12 Mar 2016 07:40:17 -0800*